better way of excluding clang-repl on iOS
This changes the build to actually not build or link clang-repl on
arm32, instead of building it and then removing it after.  It also
doesn't remove clang-repl on arm64 iOS anymore, though I haven't tested
it it might work, since I think it's CPU dependent and it probobly has
arm64 macOS support.